add keyevent to initial.py
