+++ initial.py
@@ -7,35 +7,84 @@ class Main(QWidget):
     def __init__(self, parent=None):
         QWidget.__init__(self, parent)
 
+        self.is_first_right = True
+        self.is_first_left = True
+        self.is_first_release_right = True
+        self.is_first_release_left = True
+        self.current_x = 0
+        self.current_y = 0
+
         # initial mainWindow
-        self.resize(1440, 627)  # original 1046 x 3772
+        self.resize(1440, 627)
         self.setWindowTitle("Forest of Patience")
 
         # background
         background_image = QImage('resource/back_ground_4.png')
-        modified_background_image = background_image.scaled(QSize(2000, 1080))  # original 3326 X 880
+        modified_background_image = background_image.scaled(QSize(2000, 1080))
         palette = QPalette()
         palette.setBrush(10, QBrush(modified_background_image))
         self.setPalette(palette)
-
-        # character
-        self.movie = QMovie("resource/avatar_walk1_default.gif", QByteArray(), self)
-        self.character = QLabel('asd', self)
-        character.initial(self, self.character, self.movie)
 
         # skill
         self.movie = self.movie = QMovie("resource/skill_6.gif", QByteArray(), self)
         self.skill = QLabel(self)
         skills.showSkill.initial(self, self.skill, self.movie)
 
+        # character
+        self.movie = QMovie("resource/avatar_walk1_default.gif", QByteArray(), self)
+        self.character = QLabel('asd', self)
+        character.initial(self, self.character, self.movie)
+
     def keyPressEvent(self, event):
-        character.keyPressEvent2(self, event)
+        self.current_x = self.character.pos().x()
+        self.current_y = self.character.pos().y()
+        # right
+        if event.key() == Qt.Key_D:
+            if self.is_first_right:
+                # initial character
+                self.movie = QMovie('resource/avatar_walk1_default_flip.gif', QByteArray(), self)
+                character.initial(self, self.character, self.movie)
+                self.is_first_right = False
+                self.is_first_left = True
+                self.is_first_release_right = True
+            self.character.move(self.current_x + 5, self.current_y)
+
+        # left
+        elif event.key() == Qt.Key_A:
+            if self.is_first_left:
+                # initial character
+                self.movie = QMovie('resource/avatar_walk1_default.gif', QByteArray(), self)
+                character.initial(self, self.character, self.movie)
+                self.is_first_right = True
+                self.is_first_left = False
+                self.is_first_release_left = True
+            self.character.move(self.current_x - 5, self.current_y)
 
         if event.key() == Qt.Key_Escape:
             self.close()
-    #
-    # def keyReleaseEvent(self, event):
-    #     self.character.keyReleaseEvent(event)
+
+    def keyReleaseEvent(self, event):
+        self.current_x = self.character.pos().x()
+        self.current_y = self.character.pos().y()
+        key = event.key()
+        # release right
+        if key == 68 and not event.isAutoRepeat():
+            if self.is_first_release_right:
+                self.movie = QMovie('resource/avatar_stand1_default_flip.gif', QByteArray(), self)
+                character.initial(self, self.character, self.movie)
+                self.is_first_release_right = False
+                self.is_first_right = True
+                self.character.move(self.current_x, self.current_y)
+
+        # release left
+        elif key == 65 and not event.isAutoRepeat():
+            if self.is_first_release_left:
+                self.movie = QMovie('resource/avatar_stand1_default.gif', QByteArray(), self)
+                character.initial(self, self.character, self.movie)
+                self.is_first_release_left = False
+                self.is_first_left = True
+                self.character.move(self.current_x, self.current_y)
+
 
 if __name__ == "__main__":
     app = QApplication(sys.argv)
